% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferlystat-package.R
\docType{package}
\name{inferlystat}
\alias{inferlystat-package}
\alias{inferlystat}
\title{inferlystat: Higer Level Interface for Statistical Inference}
\description{
This is a high-level package, attempting the re-imagination of "grammar of statistical inference" framework in a package, mixing with 'ggplot2' / 'tidyverse' paradigm.
}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{internal}
